add -Wconversion

+++ termite.c
@@ -114,7 +114,7 @@ static char *check_match(VteTerminal *vte, int event_x, int event_y) {
 }
 
 static gboolean button_press_cb(VteTerminal *vte, GdkEventButton *event) {
-    char *match = check_match(vte, event->x, event->y);
+    char *match = check_match(vte, (int)event->x, (int)event->y);
     if (event->button == 1 && event->type == GDK_BUTTON_PRESS && match != NULL) {
         const char *argv[] = URL_COMMAND(match);
         g_spawn_async(NULL, (char **)argv, NULL, 0, NULL, NULL, NULL, NULL);
